requirements.txt
asgiref==3.3.2
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==2.2.10
django-taggit==1.3.0
djangorestframework==3.12.4
idna==2.10
importlib-metadata==3.10.1
iniconfig==1.1.1
Markdown==3.3.4
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.3
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.7.4.3
urllib3==1.26.4
zipp==3.4.1
